Modified MainActivity
