+++ app/src/main/AndroidManifest.xml
@@ -15,7 +15,7 @@
         android:supportsRtl="true">
 
         <activity
-            android:name=".views.main.FirstActivity"
+            android:name=".views.main.MainActivity"
             android:exported="true"
             android:screenOrientation="portrait"
             android:theme="@style/Theme.ShopLiveSample"
