+++ app/src/main/java/cloud/shoplive/sample/CampaignInfo.kt
@@ -1,0 +1,3 @@
+package cloud.shoplive.sample
+
+data class CampaignInfo(val accessKey: String?, val campaignKey: String?)
