--- app/src/main/java/cloud/shoplive/sample/DeeplinkInfo.kt
@@ -1,3 +1,0 @@
-package cloud.shoplive.sample
-
-data class DeeplinkInfo(val accessKey: String, val campaignKey: String)
