+++ app/src/main/java/cloud/shoplive/sample/SchemeActivity.kt
@@ -2,7 +2,7 @@ package cloud.shoplive.sample
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
-import cloud.shoplive.sample.views.main.FirstActivity
+import cloud.shoplive.sample.views.main.MainActivity
 
 class SchemeActivity : AppCompatActivity() {
 
@@ -21,7 +21,7 @@ class SchemeActivity : AppCompatActivity() {
             intent.putExtra("campaignKey", campaignKey)
             startActivity(intent)
             */
-            startActivity(FirstActivity.buildIntentFromDeeplink(this, accessKey, campaignKey))
+            startActivity(MainActivity.buildIntentFromDeeplink(this, accessKey, campaignKey))
             finish()
         }
     }
